+++ Assets/Scripts/Lamps/Lamp.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Net;
+using Newtonsoft.Json;
 using UnityEngine;
 using VoyagerApp.Dmx;
+using VoyagerApp.Networking;
 using VoyagerApp.Utilities;
 using VoyagerApp.Videos;
 using VoyagerApp.Workspace.Views;
@@ -20,12 +22,12 @@ namespace VoyagerApp.Lamps
 
         public bool connected => (TimeUtils.Epoch - lastMessage) < TIMEOUT;
         public double lastMessage;
+        public abstract int pixels { get; }
 
-        public virtual int pixelsCount { get; }
-
+        public Itsh itsh;
         public Video video;
-        public Itsh itsh                    = Itsh.white;
-        public VideoBuffer videoBuffer      = new VideoBuffer();
+        public VideoPosition mapping;
+        public VideoBuffer buffer;
 
         internal virtual void Update(object data)
         {
@@ -38,9 +40,17 @@ namespace VoyagerApp.Lamps
             lastMessage = TimeUtils.Epoch;
         }
 
-        public virtual void SetItsh(Itsh itsh) { this.itsh = itsh; }
-        public virtual void SetVideo(Video video) { this.video = video; }
-        public abstract void PushFrame(Color32[] colors, long frame);
+        public virtual void SetItsh(Itsh itsh) => this.itsh = itsh;
+
+        public virtual void SetVideo(Video video) => this.video = video;
+
+        public virtual void PushFrame(Color32[] colors, long frame)
+        {
+            if (!buffer.FrameExists(frame))
+                buffer.SetFrame(frame, ColorUtils.ColorsToBytes(colors));
+        }
+
+        public virtual void SetMapping(VideoPosition mapping) => this.mapping = mapping;
 
         public abstract LampItemView AddToWorkspace();
         public abstract LampItemView AddToWorkspace(Vector2 position);