+++ Assets/Scripts/UI/Menus/Inspector/LoadMenu.cs
@@ -1,50 +1,46 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
-using VoyagerApp.Utilities;
+using VoyagerApp.Projects;
 using VoyagerApp.Videos;
 
 namespace VoyagerApp.UI.Menus
 {
     public class LoadMenu : Menu
     {
-        [SerializeField] LoadMenuItem itemPrefab    = null;
-        [SerializeField] Transform container        = null;
-        [SerializeField] VideoMapper mapper         = null;
-        [SerializeField] bool wm                    = false;
+        [SerializeField] LoadMenuItem itemPrefab = null;
+        [SerializeField] Transform container     = null;
 
         List<LoadMenuItem> items = new List<LoadMenuItem>();
 
         internal override void OnShow()
         {
-            foreach (var item in new List<LoadMenuItem>(items))
-            {
-                items.Remove(item);
-                Destroy(item.gameObject);
-            }
-
-            string[] saves = null;
-            if (wm)
-            {
-                if (!Directory.Exists(FileUtils.WorkspaceSavesPath + "/vm"))
-                    Directory.CreateDirectory(FileUtils.WorkspaceSavesPath + "/vm");
-                saves = Directory.GetFiles(FileUtils.WorkspaceSavesPath + "/vm", "*.ws");
-            }
-            else
-                saves = Directory.GetFiles(FileUtils.WorkspaceSavesPath, "*.ws");
-
-            foreach (var save in saves)
-            {
-                LoadMenuItem item = Instantiate(itemPrefab, container);
-                item.SetPath(save, wm);
-                items.Add(item);
-            }
+            ClearOldItems();
+            DisplayAllItems();
         }
 
-        internal void VideoMappingLoaded()
+        void ClearOldItems()
         {
-            var lamps = WorkspaceUtils.LampItems;
-            if (lamps.Count > 0) mapper.SetVideo(lamps[0].lamp.video);
+            new List<LoadMenuItem>(items).ForEach(RemoveItem);
+        }
+
+        public void RemoveItem(LoadMenuItem item)
+        {
+            items.Remove(item);
+            item.Remove();
+        }
+
+        void DisplayAllItems()
+        {
+            var projPath = Project.ProjectsDirectory;
+            var projects = Directory.GetDirectories(projPath);
+
+            foreach (var project in projects)
+            {
+                LoadMenuItem item = Instantiate(itemPrefab, container);
+                item.SetPath(project);
+                items.Add(item);
+            }
         }
     }
 }