+++ Assets/Scripts/Utilities/TimeUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using VoyagerApp.Videos;
 
 namespace VoyagerApp.Utilities
 {
@@ -7,6 +8,13 @@ namespace VoyagerApp.Utilities
         public static double Epoch
         {
             get => (DateTime.UtcNow - new DateTime(1970, 1, 1)).TotalSeconds;
+        }
+
+        public static long GetFrameOfVideo(Video video)
+        {
+            var since = Epoch - video.lastStartTime;
+            var frames = (long)(since * video.fps);
+            return frames % video.frames;
         }
     }
 }