+++ Assets/Scripts/Utilities/VectorUtils.cs
@@ -47,5 +47,32 @@ namespace VoyagerApp.Utilities
 
             return points;
         }
+
+        public static Vector2 ScreenSizeWorldSpace
+        {
+            get
+            {
+                var min = new Vector2(0, 0);
+                var max = new Vector2(Screen.width, Screen.height);
+
+                Camera cam = Camera.main;
+
+                var wMin = cam.ScreenToWorldPoint(min);
+                var wMax = cam.ScreenToWorldPoint(max);
+
+                float width = Mathf.Abs(wMax.x - wMin.x);
+                float height = Mathf.Abs(wMax.y - wMin.y);
+
+                return new Vector2(width, height);
+            }
+        }
+
+        public static float AngleFromTo(Vector2 from, Vector2 to)
+        {
+            Vector2 direction = to - from;
+            float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
+            if (angle < 0f) angle += 360f;
+            return angle;
+        }
     }
 }